完成列队，开始推送

+++ apps/user/event/Notice.php
@@ -1,26 +1,34 @@
 <?php
 namespace app\User\event;
 use app\User\model\userNotice;
-class Notice{
+class Notice extends \base\Event{
 	protected $module='User';
+	public $model='\\app\\User\\model\\userNotice';
 	public $errCode='';
 	
 	public function add($data,$autotxt=true,$autoimg=true,$download=false,$userID=false){
+		if(isset($data['id']))unset($data['id']);
+		$this->save($data,$autotxt,$autoimg,$download,$userID);
+	}
+	
+	public function save($data,$autotxt=true,$autoimg=true,$download=false,$userID=false){
 		if(!isset($data['name']) or !$data['name']){
 			$this->errCode='00010009';
 			return false;
 		}
-		if(!isset($data['content']) or !$data['content']){
-			$this->errCode='00010008';
-			return false;
-		}
-		$content=$data['content'];
-		unset($data['content']);
-		if($autotxt){
-			$data['txts']=get_html_txts($content,500);
+		if(isset($data['content'])){
+			$content=trim($data['content']);
+			if($autotxt){
+				$data['txts']=get_html_txts($content,500);
+			}else{
+				$data['txts']=isset($data['txts'])?$data['txts']:'';
+			}
 		}else{
-			$data['txts']=isset($data['txts'])?$data['txts']:'';
+			unset($data['txts']);
 		}
+		unset($data['content']);
+		
+		
 		if(!isset($data['img']) or !$data['img']){
 			if($autoimg){
 				$imgArr=get_html_img($content);
@@ -31,30 +39,35 @@ class Notice{
 				$data['img']='';
 			}
 		}
-		if(!$data['publish_time'])$data['publish_time']=get_now_time();
-		if(!$data['expire_time'])$data['expire_time']=0;
-		$data['user_id']=$userID;
-		
-		$notice=userNotice::create($data);
-		//$notice=['id'=>1];
+		if(!isset($data['id'])){
+			if(!$data['publish_time'])$data['publish_time']=get_now_time();
+			if(!$data['expire_time'])$data['expire_time']=0;
+		}
+		if($userID)$data['user_id']=$userID;
+		if(isset($data['id'])){
+			$notice=$this->model::get($data['id']);
+		}else{
+			$notice=$this->model::create($data);
+		}
 		if(!$notice){
 			$this->errCode='00010002';
 			return false;
 		}
-		$event=controller('File/Content','event');
-		$back=$event->save($content,$this->module,'notice',$notice['id'],$download);
-		if(!$back){
-			$notice->delete();
-			$this->errCode=set_err_back($event->errCode,'File');
-			return false;
+		if(isset($content)){
+			$event=controller('File/Content','event');
+			$back=$event->save($content,$this->module,'notice',$notice['id'],$download);
+			if(!$back){
+				if(!isset($data['id']))$notice->delete();
+				$this->errCode=set_err_back($event->errCode,'File');
+				return false;
+			}
+		}else{
+			$back=0;
 		}
-		$notice->content_id=$back;
-		$notice->save();
+		$dbData=isset($data['id'])?$data:['id'=>$notice['id']];
+		$dbData['content_id']=$back;
+		//$notice->content_id=$back;
+		$notice->save($dbData);
 		return $notice;
 	}
-	
-	public function getLists($wh=null,$tmp_field=true,$order=null,$page=null,$del=false){
-		
-	}
-	
 }
